requirements.txt
pandas
psycopg2-binary
sqlalchemy
scikit-learn
streamlit
python-dotenv

